core/requirements.txt
configparser==6.0.0
charset-normalizer==3.3.2
get-mac==0.9.2
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
netifaces==0.11.0
PyYAML==6.0.1
requests==2.31.0
smmap==5.0.1
pyyaml==6.0.1